Many improvements to director interface.
Moves highlight of current shot to separate class to prevent code
duplication.
